requirements.txt
ipykernel
matplotlib
slack_bolt
python-dotenv
snakemq
wheel